refactor: api 구조, 네이밍 수정
