+++ src/apis/queryHooks/basicAuction/useGetBasicAuction.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 
-import { getBasicAuction } from '@/apis/queryFunctions/basicAuctionQueryFn';
+import { getBasicAuction } from '@/apis/queryFunctions/basicAuction';
 
 function useGetBasicAuction(id: number) {
   const { data } = useQuery({
