+++ src/apis/queryHooks/basicAuction/useGetBasicAuctionList.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
 
-import { getBasicAuctionList } from '@/apis/queryFunctions/basicAuctionQueryFn';
-import { GetBasicAuctionListParams } from '@/apis/types/Auction';
+import { getBasicAuctionList } from '@/apis/queryFunctions/basicAuction';
+import { GetBasicAuctionListParams } from '@/apis/types/basicAuction';
 
 function useGetBasicAuctionList(params: GetBasicAuctionListParams) {
   const { data } = useQuery({
