+++ src/apis/queryHooks/useAuth/usePostLogin.ts
@@ -3,7 +3,7 @@ import { useMutation } from '@tanstack/react-query';
 import { postLogin } from '@/apis/queryFunctions/Auth';
 import { LoginParams } from '@/apis/types/Auth';
 
-export default function usePostLogin() {
+function usePostLogin() {
   const { mutate, error } = useMutation({
     mutationFn: (param: LoginParams) => postLogin(param),
     onSuccess: () => {
@@ -16,3 +16,5 @@ export default function usePostLogin() {
 
   return { mutate, error };
 }
+
+export default usePostLogin;
